Move function correct file and rename name stylecomponent
